test: Fix format specifiers for cross-platform compatibility
Use (unsigned int) casts for sizeof and OpenXR uint32_t types instead
of %zu and bare %u to ensure compatibility with all compilers.

+++ test/testgpu_spinning_cube_xr.c
@@ -518,7 +518,7 @@ static bool create_cube_buffers(void)
     SDL_SubmitGPUCommandBuffer(cmd);
     SDL_ReleaseGPUTransferBuffer(gpu_device, transfer);
 
-    SDL_Log("Created cube vertex (%zu bytes) and index (%zu bytes) buffers", sizeof(vertices), sizeof(indices));
+    SDL_Log("Created cube vertex (%u bytes) and index (%u bytes) buffers", (unsigned int)sizeof(vertices), (unsigned int)sizeof(indices));
     return true;
 }
 
@@ -606,8 +606,8 @@ static bool create_swapchains(void)
         xr_views[i].pose.orientation.w = 1.0f;
 
         SDL_Log("Eye %" SDL_PRIu32 ": recommended %ux%u", i,
-                view_configs[i].recommendedImageRectWidth,
-                view_configs[i].recommendedImageRectHeight);
+                (unsigned int)view_configs[i].recommendedImageRectWidth,
+                (unsigned int)view_configs[i].recommendedImageRectHeight);
 
         /* Create swapchain using OpenXR's XrSwapchainCreateInfo */
         XrSwapchainCreateInfo swapchain_info = { XR_TYPE_SWAPCHAIN_CREATE_INFO };
